Implement DataTransfer for disabled MPI support
