+++ source/distributed/tria.cc
@@ -1868,13 +1868,13 @@ namespace parallel
       const typename dealii::internal::p4est::types<dim>::gloidx
         *previous_global_first_quadrant)
     {
-      Assert(this->data_transfer->sizes_fixed_cumulative.size() > 0,
+      Assert(this->data_transfer.sizes_fixed_cumulative.size() > 0,
              ExcMessage("No data has been packed!"));
 
       // Resize memory according to the data that we will receive.
-      this->data_transfer->dest_data_fixed.resize(
+      this->data_transfer.dest_data_fixed.resize(
         parallel_forest->local_num_quadrants *
-        this->data_transfer->sizes_fixed_cumulative.back());
+        this->data_transfer.sizes_fixed_cumulative.back());
 
       // Execute non-blocking fixed size transfer.
       typename dealii::internal::p4est::types<dim>::transfer_context
@@ -1885,14 +1885,14 @@ namespace parallel
           previous_global_first_quadrant,
           parallel_forest->mpicomm,
           0,
-          this->data_transfer->dest_data_fixed.data(),
-          this->data_transfer->src_data_fixed.data(),
-          this->data_transfer->sizes_fixed_cumulative.back());
+          this->data_transfer.dest_data_fixed.data(),
+          this->data_transfer.src_data_fixed.data(),
+          this->data_transfer.sizes_fixed_cumulative.back());
 
-      if (this->data_transfer->variable_size_data_stored)
+      if (this->data_transfer.variable_size_data_stored)
         {
           // Resize memory according to the data that we will receive.
-          this->data_transfer->dest_sizes_variable.resize(
+          this->data_transfer.dest_sizes_variable.resize(
             parallel_forest->local_num_quadrants);
 
           // Execute fixed size transfer of data sizes for variable size
@@ -1902,23 +1902,23 @@ namespace parallel
             previous_global_first_quadrant,
             parallel_forest->mpicomm,
             1,
-            this->data_transfer->dest_sizes_variable.data(),
-            this->data_transfer->src_sizes_variable.data(),
+            this->data_transfer.dest_sizes_variable.data(),
+            this->data_transfer.src_sizes_variable.data(),
             sizeof(unsigned int));
         }
 
       dealii::internal::p4est::functions<dim>::transfer_fixed_end(tf_context);
 
       // Release memory of previously packed data.
-      this->data_transfer->src_data_fixed.clear();
-      this->data_transfer->src_data_fixed.shrink_to_fit();
+      this->data_transfer.src_data_fixed.clear();
+      this->data_transfer.src_data_fixed.shrink_to_fit();
 
-      if (this->data_transfer->variable_size_data_stored)
+      if (this->data_transfer.variable_size_data_stored)
         {
           // Resize memory according to the data that we will receive.
-          this->data_transfer->dest_data_variable.resize(
-            std::accumulate(this->data_transfer->dest_sizes_variable.begin(),
-                            this->data_transfer->dest_sizes_variable.end(),
+          this->data_transfer.dest_data_variable.resize(
+            std::accumulate(this->data_transfer.dest_sizes_variable.begin(),
+                            this->data_transfer.dest_sizes_variable.end(),
                             std::vector<int>::size_type(0)));
 
 #  if DEAL_II_P4EST_VERSION_GTE(2, 0, 65, 0)
@@ -1928,10 +1928,10 @@ namespace parallel
           // at all, which is mandatory if one of our processes does not own
           // any quadrant. This bypasses the assertion from being triggered.
           //   - see: https://github.com/cburstedde/p4est/issues/48
-          if (this->data_transfer->src_sizes_variable.size() == 0)
-            this->data_transfer->src_sizes_variable.resize(1);
-          if (this->data_transfer->dest_sizes_variable.size() == 0)
-            this->data_transfer->dest_sizes_variable.resize(1);
+          if (this->data_transfer.src_sizes_variable.size() == 0)
+            this->data_transfer.src_sizes_variable.resize(1);
+          if (this->data_transfer.dest_sizes_variable.size() == 0)
+            this->data_transfer.dest_sizes_variable.resize(1);
 #  endif
 
           // Execute variable size transfer.
@@ -1940,16 +1940,16 @@ namespace parallel
             previous_global_first_quadrant,
             parallel_forest->mpicomm,
             1,
-            this->data_transfer->dest_data_variable.data(),
-            this->data_transfer->dest_sizes_variable.data(),
-            this->data_transfer->src_data_variable.data(),
-            this->data_transfer->src_sizes_variable.data());
+            this->data_transfer.dest_data_variable.data(),
+            this->data_transfer.dest_sizes_variable.data(),
+            this->data_transfer.src_data_variable.data(),
+            this->data_transfer.src_sizes_variable.data());
 
           // Release memory of previously packed data.
-          this->data_transfer->src_sizes_variable.clear();
-          this->data_transfer->src_sizes_variable.shrink_to_fit();
-          this->data_transfer->src_data_variable.clear();
-          this->data_transfer->src_data_variable.shrink_to_fit();
+          this->data_transfer.src_sizes_variable.clear();
+          this->data_transfer.src_sizes_variable.shrink_to_fit();
+          this->data_transfer.src_data_variable.clear();
+          this->data_transfer.src_data_variable.shrink_to_fit();
         }
     }
 
@@ -3385,7 +3385,7 @@ namespace parallel
       // pack data before triangulation gets updated
       if (this->cell_attached_data.n_attached_data_sets > 0)
         {
-          this->data_transfer->pack_data(
+          this->data_transfer.pack_data(
             this->local_cell_relations,
             this->cell_attached_data.pack_callbacks_fixed,
             this->cell_attached_data.pack_callbacks_variable);
@@ -3418,7 +3418,7 @@ namespace parallel
                                  previous_global_first_quadrant.data());
 
           // also update the CellStatus information on the new mesh
-          this->data_transfer->unpack_cell_status(this->local_cell_relations);
+          this->data_transfer.unpack_cell_status(this->local_cell_relations);
         }
 
 #  ifdef DEBUG
@@ -3558,7 +3558,7 @@ namespace parallel
       // pack data before triangulation gets updated
       if (this->cell_attached_data.n_attached_data_sets > 0)
         {
-          this->data_transfer->pack_data(
+          this->data_transfer.pack_data(
             this->local_cell_relations,
             this->cell_attached_data.pack_callbacks_fixed,
             this->cell_attached_data.pack_callbacks_variable);
