+++ source/grid/CMakeLists.txt
@@ -32,6 +32,7 @@ endif()
 
 set(_unity_include_src
   cell_id.cc
+  data_transfer.cc
   grid_refinement.cc
   intergrid_map.cc
   manifold.cc
@@ -70,6 +71,7 @@ setup_source_list("${_unity_include_src}"
 
 set(_inst
   cell_id.inst.in
+  data_transfer.inst.in
   grid_generator_cgal.inst.in
   grid_generator_from_name.inst.in
   grid_generator.inst.in
